backend/requirements.txt
fastapi
uvicorn
pandas
langgraph
langchain-groq
python-multipart
pydantic-settings
duckdb
openpyxl
langgraph-checkpoint-sqlite
